requirements.txt
requests
flask
discord.py
gitpython
google-api-python-client
openai
